Rename package directories: part 1 (due to git misbehaviour)
