+++ CondCore/SiPhase2TrackerPlugins/src/plugin.cc
@@ -1,5 +1,5 @@
 #include "CondCore/PluginSystem/interface/registration_macros.h"
-#include "CondFormats/Phase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
 #include "CondFormats/DataRecord/interface/TrackerDetToDTCELinkCablingMapRcd.h"
 
 REGISTER_PLUGIN(TrackerDetToDTCELinkCablingMapRcd,TrackerDetToDTCELinkCablingMap);
