+++ CondFormats/SiPhase2TrackerObjects/src/T_EventSetup_TrackerDetToDTCELinkCablingMap.cc
@@ -1,4 +1,4 @@
-#include "CondFormats/Phase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
 #include "FWCore/Utilities/interface/typelookup.h"
 
 TYPELOOKUP_DATA_REG(TrackerDetToDTCELinkCablingMap);
