+++ CondFormats/SiPhase2TrackerObjects/src/classes.h
@@ -1,5 +1,5 @@
-#include "CondFormats/Phase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
-#include "CondFormats/Phase2TrackerObjects/interface/DTCELinkId.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/DTCELinkId.h"
 #include "DataFormats/DetId/interface/DetId.h"
 
 namespace
