+++ CondFormats/SiPhase2TrackerObjects/src/headers.h
@@ -1,0 +1,2 @@
+#include "CondFormats/SiPhase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/DTCELinkId.h"
