+++ CondTools/SiPhase2Tracker/BuildFile.xml
@@ -3,7 +3,7 @@
 <use name="FWCore/ServiceRegistry"/>
 <use name="CondCore/DBOutputService"/>
 <use name="CondFormats/Common"/>
-<use name="CondFormats/Phase2TrackerObjects"/>
+<use name="CondFormats/SiPhase2TrackerObjects"/>
 <use name="CondFormats/DataRecord"/>
 <use name="rootrflx"/>
 <!--<use name="boost"/> -->
