+++ CondTools/SiPhase2Tracker/plugins/DTCCablingMapProducer.cc
@@ -35,8 +35,8 @@ Implementation:
 
 #include "CondFormats/Common/interface/Time.h"
 #include "CondFormats/DataRecord/interface/TrackerDetToDTCELinkCablingMapRcd.h"
-#include "CondFormats/Phase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
-#include "CondFormats/Phase2TrackerObjects/interface/DTCELinkId.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/TrackerDetToDTCELinkCablingMap.h"
+#include "CondFormats/SiPhase2TrackerObjects/interface/DTCELinkId.h"
